requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
PyMySQL
Werkzeug
python-dotenv
psycopg2-binary